new payment list and payment details screens.

+++ app/src/main/java/dev/nelson/mot/main/data/room/model/payment/PaymentEntity.kt
@@ -17,10 +17,10 @@ data class PaymentEntity(
     @ColumnInfo(name = PaymentTable.TITLE) var title: String,
     @ColumnInfo(name = PaymentTable.COST) val cost: Int,
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = PaymentTable.ID) val id: Int? = null,
+    @ColumnInfo(name = PaymentTable.ID) val id: Long? = null,
     @ColumnInfo(name = PaymentTable.SUMMARY) var summary: String? = null,
     @ColumnInfo(name = PaymentTable.CATEGORY_ID_KEY) val categoryIdKey: Int? = null,
     @ColumnInfo(name = PaymentTable.DATE) val date: String? = null,
-    @ColumnInfo(name = PaymentTable.DATE_IN_MILLISECONDS) val dateInMilliseconds: Int? = null
+    @ColumnInfo(name = PaymentTable.DATE_IN_MILLISECONDS) val dateInMilliseconds: Long? = null
 )
